Project rename and synchronization init attempts

+++ FileSynchro/File.cs
@@ -1,11 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
+using System.Security.Cryptography;
 using System.Text;
 using System.Threading.Tasks;
 
 namespace FileSync
-{
+{ 
     public class File
     {
         public int FileId { get; set; }
@@ -15,7 +17,7 @@ namespace FileSync
         public DateTime? FileLastModificationDate { get; set; }
         public DateTime? FilePlacementDate { get; set; }
         public string FileType { get; set; }
-        public int FileSize { get; set; }
+        public long FileSize { get; set; }
         public string FileLocation { get; set; }
     }
 }
